refactor(utils, config): split into logging and error_handling modules
